Extract resource/resources from metadata
This extracts resource of the PID from the metadata if a given
PID supports providing the resource in the PID metadata. If there
is only a single resource the request is redirected to the given
resource and if there is more than one resource a list of resources
is displayed in order to choose from.